# evc-app/src/_assets/api.yml
swagger: "2.0"
info:
  description: AUA backend API
  version: "1.0.0"
  title: "AUA API"
basePath: "/api/v1"
schemes:
  - http
  - https
...
paths:
...
  /auth/reset_password/{token}:
    get:
      operationId: retrievePassword
      # security:
      #   - authAnyRole: []
      parameters:
        - name: token
          in: path
          type: string
          format: uuid
          required: true
      responses:
        200:
          description: OK
...